Rename _loggers.py to log.py
