+++ cashocs/__init__.py
@@ -31,8 +31,6 @@ from cashocs._constraints.constrained_problems import (
 )
 from cashocs._constraints.constraints import EqualityConstraint
 from cashocs._constraints.constraints import InequalityConstraint
-from cashocs._loggers import LogLevel
-from cashocs._loggers import set_log_level
 from cashocs._optimization import verification
 from cashocs._optimization.cost_functional import Functional
 from cashocs._optimization.cost_functional import IntegralFunctional
@@ -55,6 +53,8 @@ from cashocs.geometry import regular_mesh
 from cashocs.io import convert
 from cashocs.io import import_mesh
 from cashocs.io import load_config
+from cashocs.log import LogLevel
+from cashocs.log import set_log_level
 from cashocs.nonlinear_solvers import linear_solve
 from cashocs.nonlinear_solvers import newton_solve
 from cashocs.nonlinear_solvers import picard_iteration
