+++ cashocs/_optimization/line_search/polynomial_line_search.py
@@ -28,7 +28,7 @@ import numpy as np
 from typing_extensions import TYPE_CHECKING
 
 from cashocs import _exceptions
-from cashocs import _loggers
+from cashocs import log
 from cashocs._optimization.line_search import line_search
 
 if TYPE_CHECKING:
@@ -84,7 +84,7 @@ class PolynomialLineSearch(line_search.LineSearch):
             return True
 
         if self.stepsize * self.search_direction_inf <= 1e-8:
-            _loggers.error("Stepsize too small.")
+            log.error("Stepsize too small.")
             solver.line_search_broken = True
             return True
         elif (
@@ -92,7 +92,7 @@ class PolynomialLineSearch(line_search.LineSearch):
             and not self.is_newton
             and self.stepsize / self.armijo_stepsize_initial <= 1e-8
         ):
-            _loggers.error("Stepsize too small.")
+            log.error("Stepsize too small.")
             solver.line_search_broken = True
             return True
 
@@ -161,7 +161,7 @@ class PolynomialLineSearch(line_search.LineSearch):
             objective_step = self.cost_functional.evaluate()
             self.f_vals.append(objective_step)
 
-            _loggers.debug(
+            log.debug(
                 f"Line search - Trial stepsize {self.stepsize:.3e} - "
                 f"Function value {objective_step:.3e}"
             )
@@ -171,9 +171,9 @@ class PolynomialLineSearch(line_search.LineSearch):
             if self._satisfies_armijo_condition(
                 objective_step, current_function_value, decrease_measure
             ):
-                _loggers.debug("Stepsize satisfies the Armijo decrease condition.")
+                log.debug("Stepsize satisfies the Armijo decrease condition.")
                 if self.optimization_variable_abstractions.requires_remeshing():
-                    _loggers.debug(
+                    log.debug(
                         "The mesh quality was sufficient for accepting the step, "
                         "but the mesh cannot be used anymore for computing a gradient."
                         "Performing a remeshing operation."
@@ -190,9 +190,7 @@ class PolynomialLineSearch(line_search.LineSearch):
                 break
 
             else:
-                _loggers.debug(
-                    "Stepsize does not satisfy the Armijo decrease condition."
-                )
+                log.debug("Stepsize does not satisfy the Armijo decrease condition.")
                 self.stepsize = self._compute_polynomial_stepsize(
                     current_function_value,
                     decrease_measure,
