+++ cashocs/_optimization/shape_optimization/shape_optimization_problem.py
@@ -34,11 +34,11 @@ except ImportError:
 
 from cashocs import _exceptions
 from cashocs import _forms
-from cashocs import _loggers
 from cashocs import _pde_problems
 from cashocs import _utils
 from cashocs import geometry
 from cashocs import io
+from cashocs import log
 from cashocs._optimization import cost_functional
 from cashocs._optimization import line_search as ls
 from cashocs._optimization import optimization_algorithms
@@ -231,7 +231,7 @@ class ShapeOptimizationProblem(optimization_problem.OptimizationProblem):
         if self.uses_custom_scalar_product and self.config.getboolean(
             "ShapeGradient", "use_p_laplacian"
         ):
-            _loggers.warning(
+            log.warning(
                 (
                     "You have supplied a custom scalar product and set the parameter "
                     "``use_p_laplacian`` in the config file."
@@ -480,7 +480,7 @@ class ShapeOptimizationProblem(optimization_problem.OptimizationProblem):
         self.solver.post_processing()
 
     def _clear_remesh_directory(self) -> None:
-        _loggers.debug("An exception was raised, deleting the created temporary files.")
+        log.debug("An exception was raised, deleting the created temporary files.")
         if (
             not self.config.getboolean("Debug", "remeshing")
             and fenics.MPI.rank(fenics.MPI.comm_world) == 0
