+++ cashocs/_pde_problems/hessian_problems.py
@@ -28,8 +28,8 @@ from typing import List, TYPE_CHECKING
 import fenics
 import numpy as np
 
-from cashocs import _loggers
 from cashocs import _utils
+from cashocs import log
 from cashocs import nonlinear_solvers
 
 if TYPE_CHECKING:
@@ -324,7 +324,7 @@ class HessianProblem:
 
             rsnew = self.form_handler.scalar_product(self.residual, self.residual)
             eps = np.sqrt(rsnew)
-            _loggers.debug(f"Residual of the CG method: {eps/eps_0:.3e} (relative)")
+            log.debug(f"Residual of the CG method: {eps/eps_0:.3e} (relative)")
             if eps < self.inner_newton_atol + self.inner_newton_rtol * eps_0:
                 break
 
@@ -387,7 +387,7 @@ class HessianProblem:
             eps = np.sqrt(
                 self.form_handler.scalar_product(self.residual, self.residual)
             )
-            _loggers.debug(f"Residual of the CR method: {eps/eps_0:.3e} (relative)")
+            log.debug(f"Residual of the CR method: {eps/eps_0:.3e} (relative)")
             if (
                 eps < self.inner_newton_atol + self.inner_newton_rtol * eps_0
                 or i == self.max_it_inner_newton - 1
