+++ cashocs/geometry/mesh_handler.py
@@ -30,9 +30,9 @@ import fenics
 import numpy as np
 
 from cashocs import _exceptions
-from cashocs import _loggers
 from cashocs import _utils
 from cashocs import io
+from cashocs import log
 from cashocs._optimization import line_search as ls
 from cashocs.geometry import deformations
 from cashocs.geometry import quality
@@ -156,7 +156,7 @@ class _MeshHandler:
         )
 
         if self.mesh_quality_tol_lower > 0.9 * self.mesh_quality_tol_upper:
-            _loggers.warning(
+            log.warning(
                 "You are using a lower remesh tolerance (tol_lower) close to the upper "
                 "one (tol_upper). This may slow down the optimization considerably."
             )
@@ -302,7 +302,7 @@ class _MeshHandler:
             raise _exceptions.CashocsException("Not a valid mesh transformation")
 
         if not self.a_priori_tester.test(transformation, self.volume_change):
-            _loggers.debug("Mesh transformation rejected due to a priori check.")
+            log.debug("Mesh transformation rejected due to a priori check.")
             return False
         else:
             success_flag = self.deformation_handler.move_mesh(
@@ -652,7 +652,7 @@ class _MeshHandler:
         mesh_quality_tol_upper = self.db.config.getfloat("MeshQuality", "tol_upper")
 
         if mesh_quality_tol_lower > 0.9 * mesh_quality_tol_upper:
-            _loggers.warning(
+            log.warning(
                 "You are using a lower remesh tolerance (tol_lower) close to "
                 "the upper one (tol_upper). This may slow down the "
                 "optimization considerably."
@@ -683,11 +683,11 @@ class _MeshHandler:
         """
         if self.config.getboolean("ShapeGradient", "global_deformation"):
             pre_log_level = (
-                _loggers._cashocs_logger.level  # pylint: disable=protected-access
+                log._cashocs_logger.level  # pylint: disable=protected-access
             )
-            _loggers.set_log_level(_loggers.LogLevel.WARNING)
+            log.set_log_level(log.LogLevel.WARNING)
             mesh, _, _, _, _, _ = import_mesh(xdmf_filename)
-            _loggers.set_log_level(pre_log_level)
+            log.set_log_level(pre_log_level)
 
             deformation_space = fenics.VectorFunctionSpace(mesh, "CG", 1)
             interpolator = _utils.Interpolator(
